Create pages directory & move home to pages
