+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { Route, Switch } from 'react-router-dom'
 import Header from './components/Header'
 import Footer from './components/Footer'
-import Home from './components/Home'
+import Home from './pages/Home'
 
 const App = () => {
   return (
